chore: :truck: Rename ValueSequence folder -> Sequence
