+++ Packages/com.careboo.blinq/CareBoo.Blinq/Sequence/Sequence.All.cs
@@ -1,0 +1,21 @@
+﻿namespace CareBoo.Blinq
+{
+    public static partial class Sequence
+    {
+        public static bool All<T, TSource, TPredicate>(
+            this ref ValueSequence<T, TSource> source,
+            ValueFunc<T, bool>.Impl<TPredicate> predicate
+            )
+            where T : struct
+            where TSource : struct, ISequence<T>
+            where TPredicate : struct, IFunc<T, bool>
+        {
+            var sourceList = source.Execute();
+            for (var i = 0; i < sourceList.Length; i++)
+                if (!predicate.Invoke(sourceList[i]))
+                    return false;
+            return true;
+        }
+
+    }
+}
